+++ CHANGELOG.md
@@ -29,6 +29,7 @@ for now removed features.
 
 - **select**: autocomplete filter is now updated when an option is selected
 - **tabs**: tabs can be dynamically removed
+- **grid-pagination**: improve accessibility
 
 ### Security